test: Add comprehensive CodeRabbit nits detection integration tests
This commit adds extensive integration tests for Issue #121 fix:

- TestCodeRabbitNitsDetectionIntegration: Tests various CodeRabbit structured comment formats
- TestCodeRabbitPatternsBackwardCompatibility: Ensures existing patterns still work
- TestActualPR120CodeRabbitReview: Tests with real PR #120 CodeRabbit review data

Key test coverage:
- CodeRabbit's structured HTML format with <details> and <summary> tags
- Emoji-based nitpick indicators (🧹 Nitpick comments)
- Mixed traditional and structured patterns
- Backward compatibility with existing nit: patterns
- Real-world CodeRabbit review format validation

All tests pass, confirming that CodeRabbit bot reviews with structured
nitpick comments are now properly detected and converted to low-priority tasks.
